[ECS] added scene submodule + material cleanup
Added a very simple scene that can be used to tie entities together.
need to investigate the best way to do a transform hierarchy using the flecs library.

By adding a scene lots of the code that was cluttering up the Renderer has been removed and is managed via the Scene & resource manager.

Did a minor clean up in the renderer update function.

a header refactor is really in order for renderer.h it's become pretty horrid when trying to use it outside of the renderer class, thinking of removing all of the glue style functions and instead providing a static getter for device. buffer, swapchain ,. . ect similar to that of ResourceManager, Scene ect. there is a static getter for the renderer class itself but currently only provides access to the functions and not the actual components that renderer houses.

overall flecs was nice and simple to integrate into the engine :^)