[GFX] Renderer header housekeeping.
Decided to clean up the header file for the Renderer it was only initially meant to be glue for a few functions but I kept on adding more, so I have replaced all of the renderer glue style functions with calls to the renderer Vulkan interface classes.

There is still more clean up to be done in a header such as moving out all of the helper functions into a potential gfx utility namespace that acts as a interface for the buffer functions, alternatively we could just exposing the VMA & highly useful portions of the vulkanBuffer class as static. this may also be a reasonable solution. :^)