[GFX] Renderer header housekeeping cont.
Continued renderer header house clean up, moved includes out of header file and replaced with forward declaration.

also removed passing in the Engine as a reference to the renderer. as a result I added some utility static functions to the window class to make it easier to get the window size/aspect/framebuffer_size. no real functionality has changed here but it's cleaner and faster to compile, I was winding whenever I changed the "backed" code I was recompiling more files than I expected as the renderer header is included in large number of places.

There is still some clean up to do relating to the window / input manager. I want to separate out the input manager from being inside of the Window class. the input manager will still use GLFW on the backend as about 60% of the functionality here but if memory serves from the last engine some input types didn't behave as expected.

I've not moved Mesh/Texture out of the renderer as they need to be moved to  a location that has yet to be decided :^)