[GFX][ECS] Input manager, material + editor camera
Refactored the Input manager to be a subsystem, as previously it was contained inside of the windowing subsystem which was just a convenient hack.

When refactoring the input code I added a static getter to make it convenient for systems as they are contained within lambdas (ECS). alongside this I moved the keycode enums to types.h and removed gamepad support for the time being. I've removed this code as when it comes to that point I would like to add support for mapping controller/keyboard inputs to various data types in the editor.

cleaned up toggle full screen code with new input manager

I added some temp code to the material to support 5 textures and made them visible on the shader.

Added editor camera "script" that is being managed via the ECS and takes advantage of the input system changes. :^)